Create README.md
I am in a fix that how can i be a computer programmer with a bachelor degree on dental science background. But i am really very much excited to learn a new skill in this digital world. So i thought that learning programming language will be better than any other option.